Bump Cocos2D to release-2.1
[UPDATED]   Cocos2D to release-2.1

+++ Pearl-Cocos2D/View/PearlCCBarSprite.h
@@ -35,12 +35,12 @@
     float   _currentLength;
 
     CGSize     _textureSize;
-    NSUInteger _uniformColor;
+    int _uniformColor;
 }
 
 @property (nonatomic, readwrite, assign) CGPoint    target;
 @property (nonatomic, readwrite, assign) CGSize     textureSize;
-@property (nonatomic, readwrite, assign) NSUInteger uniformColor;
+@property (nonatomic, readwrite, assign) int uniformColor;
 
 
 #pragma mark ###############################
